Add filter key events to bank
