+++ dotcom-rendering/src/web/components/TopicFilterBank.stories.tsx
@@ -1,6 +1,6 @@
 import { css } from '@emotion/react';
 import { ArticleDesign, ArticleDisplay, ArticlePillar } from '@guardian/libs';
-import { TopicFilterBank } from './TopicFilterBank';
+import { TopicFilterBank } from './TopicFilterBank.importable';
 
 const availableTopics: Topic[] = [
 	{ type: 'GPE', value: 'London', count: 16 },
@@ -26,6 +26,13 @@ export default {
 	title: 'Components/TopicFilterBank',
 };
 
+const baseProperties = {
+	id: '123',
+	elements: [],
+	attributes: { keyEvent: false, pinned: false, summary: false },
+	primaryDateLine: '',
+	secondaryDateLine: '',
+};
 const Wrapper = ({ children }: { children: React.ReactNode }) => (
 	<div
 		css={css`
@@ -43,6 +50,15 @@ export const topicBank = () => {
 			<TopicFilterBank
 				availableTopics={availableTopics}
 				format={format}
+				keyEvents={[
+					{
+						...baseProperties,
+						blockFirstPublished: 1638279933000,
+						title: 'title',
+					},
+				]}
+				filterKeyEvents={false}
+				keyEventsId={'key-events-filter-mobile'}
 			/>
 		</Wrapper>
 	);
