+++ dotcom-rendering/src/web/lib/LiveBlogRenderer.tsx
@@ -6,7 +6,7 @@ import { KeyEventsCarousel } from '../components/KeyEventsCarousel.importable';
 import { LiveBlock } from '../components/LiveBlock';
 import { LiveBlogEpic } from '../components/LiveBlogEpic.importable';
 import { PinnedPost } from '../components/PinnedPost';
-import { TopicFilterBank } from '../components/TopicFilterBank';
+import { TopicFilterBank } from '../components/TopicFilterBank.importable';
 
 type Props = {
 	format: ArticleFormat;
@@ -91,12 +91,14 @@ export const LiveBlogRenderer = ({
 							format={format}
 						/>
 					</Island>
-					<Island deferUntil="visible">
-						<FilterKeyEventsToggle
-							filterKeyEvents={filterKeyEvents}
-							id="filter-toggle-mobile"
-						/>
-					</Island>
+					{!switches.automaticFilters && (
+						<Island deferUntil="visible">
+							<FilterKeyEventsToggle
+								filterKeyEvents={filterKeyEvents}
+								id="filter-toggle-mobile"
+							/>
+						</Island>
+					)}
 				</Hide>
 			) : (
 				<></>
@@ -107,6 +109,9 @@ export const LiveBlogRenderer = ({
 					<TopicFilterBank
 						availableTopics={availableTopics}
 						format={format}
+						keyEvents={keyEvents}
+						filterKeyEvents={filterKeyEvents}
+						keyEventsId={'key-events-filter-mobile'}
 					/>
 				</Hide>
 			)}
